+++ libs/coin-modules/coin-bitcoin/src/__tests__/unit/synchronisation.unit.test.ts
